Fixed not deleting from dashboard widget that is not found

+++ src/main/java/com/epam/ta/reportportal/core/dashboard/impl/UpdateDashboardHandler.java
@@ -33,7 +33,6 @@ import com.epam.ta.reportportal.database.entity.ProjectRole;
 import com.epam.ta.reportportal.database.entity.user.UserRole;
 import com.epam.ta.reportportal.database.entity.widget.Widget;
 import com.epam.ta.reportportal.events.DashboardUpdatedEvent;
-import com.epam.ta.reportportal.exception.ReportPortalException;
 import com.epam.ta.reportportal.ws.model.ErrorType;
 import com.epam.ta.reportportal.ws.model.OperationCompletionRS;
 import com.epam.ta.reportportal.ws.model.dashboard.UpdateDashboardRQ;
@@ -159,14 +158,10 @@ public class UpdateDashboardHandler implements IUpdateDashboardHandler {
 			//remove from dashboard
 			dashboard.getWidgets().removeIf(w -> w.getWidgetId().equals(it));
 			Widget widget = widgetRepository.findOneLoadACL(it);
-			try {
-				if (AclUtils.isAllowedToDeleteWidget(dashboard.getAcl(), widget.getAcl(), userName,
-						projectRoles.get(projectName), userRole)) {
-					widgetRepository.delete(it);
-				}
-			} catch (Exception e) {
-				throw new ReportPortalException("Error during deleting widget", e);
-			}
+            if (null != widget && AclUtils.isAllowedToDeleteWidget(dashboard.getAcl(), widget.getAcl(),
+                    userName, projectRoles.get(projectName), userRole)) {
+                widgetRepository.delete(it);
+            }
 		});
 
 		ofNullable(rq.getShare()).ifPresent(it -> sharingService.modifySharing(Lists.newArrayList(dashboard), userName, projectName, it));
